+++ docs/conf.py
@@ -15,12 +15,11 @@
 import datetime
 import os
 import sys
-
-import pkg_resources
+from importlib.metadata import distribution
 
 
 sys.path.append(os.path.abspath('../src'))
-rqmt = pkg_resources.require('Products.GenericSetup')[0]
+rqmt = distribution('Products.GenericSetup').version
 
 
 # If your extensions are in another directory, add it here. If the directory
@@ -53,9 +52,10 @@ copyright = '2008-%s, Zope Foundation and Contributors' % year
 # other places throughout the built documents.
 #
 # The short X.Y version.
-version = '%s.%s' % tuple(rqmt.version.split('.')[:2])
+version = '%s.%s' % tuple(rqmt.split('.')[:2])
 # The full version, including alpha/beta/rc tags.
-release = rqmt.version
+release = rqmt
+
 
 # There are two options for replacing |today|: either, you set today to some
 # non-false value, then it is used: